Remove old contact search module, and distinction between simple and complex searches
